Converted to a maven project, added log4j and junit dependencies, also added method that calculates cost of traversing path without decrementing from current power usage getPowerCost(Location starting, Location ending) in PowerManager.  Updated DoorCell to correctly display closed DoorCell using unicode.

+++ SE459_cleansweep/src/edu/cleansweep/simulator/power/ChargingPowerState.java
@@ -2,7 +2,7 @@ package edu.cleansweep.simulator.power;
 
 class ChargingPowerState extends PowerState {
 	
-	ChargingPowerState(int chargeValue){
+	ChargingPowerState(double chargeValue){
 		System.out.println("[PowerState] Entering ChargingPowerState");
 		_currentCharge = chargeValue;
 	}
